implemented document chatbot into employee page

+++ application/views/bot/document_view.php
@@ -176,7 +176,7 @@ if (empty($pdf_files)) {
                 <div class="container-fluid">
 
                     <!-- Title -->
-                    <div class="row">
+                    <div class="row <?php if ($this->session->userdata('user_role') == "Employee") {echo "pt-5";}?>">
                         <div class="col-xl-9">
                             <!-- Page Heading -->
                             <div class="d-sm-flex align-items-center justify-content-between mb-4">
@@ -185,7 +185,7 @@ if (empty($pdf_files)) {
                         </div>
                         <div class="col-xl-3">
                             <div class="d-flex justify-content-end mb-4">
-                                <a type="button" class="btn bluebtn" data-toggle="modal" data-target="#view_item">Upload Files<i class="fas fa-plus pl-2"></i></a>
+                                <a type="button" class="btn bluebtn" data-toggle="modal" data-target="#view_item"><?php if ($this->session->userdata('user_role') == "Admin") {echo 'Upload File <i class="fas fa-plus pl-2"></i>';}else{echo "View Files";}?></a>
                             </div>
                         </div>
                     </div>
@@ -202,7 +202,7 @@ if (empty($pdf_files)) {
                                     <div class="row ">
                                         <div class="col-xl-2" style="border-right: black;" id="conversation_list" style="max-height: 72vh; overflow-y: auto;">
 
-
+        
                                         </div>
 
                                         <div class="col-xl-10 px-5 pb-5 " id="conversation_body" style="max-height: 72vh; overflow-y: auto;">
@@ -276,9 +276,11 @@ if (empty($pdf_files)) {
                                             </div>
                                         </div>
                                         <div class="row p-2 pb-4">
+                                        <?php if ($this->session->userdata('user_role') == "Admin") {?>
                                             <div class="col-md-12  text-right">
                                                 <a type="button" onclick="showPdfUploadDialog()" class="btn bluebtn" style="background-color: #3b75f2; color:white;">New File<i class="fas fa-plus pl-2"></i></a>
                                             </div>
+                                            <?php } ?>
                                         </div>
                                         <div class="row" id="file_grid">
                                             <?php foreach ($pdf_files as $pdf_file) : ?>
@@ -293,7 +295,9 @@ if (empty($pdf_files)) {
                                                                 <p class="px-1" style="font-size: 0.7rem;"><?php echo date("F j, Y, g:i a", strtotime($pdf_file->upload_date)); ?></p>
                                                                 <div class="group_buttons">
                                                                     <a class="button view-button unique-view-button" style="background-color: #3b75f2; color:white;" href="<?php echo base_url('assets/files/' . $pdf_file->doc_name . '.pdf'); ?>" target="_blank">Open</a>
-                                                                    <button class="delete-button unique-delete-button mt-2" onclick="delete_file(<?= $pdf_file->doc_id ?>)" id="button<?= $pdf_file->doc_id ?>" data-id="<?php echo $pdf_file->doc_id; ?>" data-name="<?= $pdf_file->doc_name ?>">Delete</button>
+                                                                    <?php if ($this->session->userdata('user_role') == "Admin") { ?>
+                                                                        <button class="delete-button unique-delete-button mt-2" onclick="delete_file(<?= $pdf_file->doc_id ?>)" id="button<?= $pdf_file->doc_id ?>" data-id="<?php echo $pdf_file->doc_id; ?>" data-name="<?= $pdf_file->doc_name ?>">Delete</button>
+                                                                    <?php } ?>
                                                                 </div>
                                                             </div>
                                                         </div>
